Remove hardcoded path
