+++ Sources/PeripheryKit/Indexer/XibIndexer.swift
@@ -1,7 +1,7 @@
 import PathKit
 
 final class XibIndexer: TypeIndexer {
-    static func make(buildPlan: BuildPlan, graph: SourceGraph) -> Self {
+    static func make(buildPlan: BuildPlan, indexStore: IndexStore, graph: SourceGraph) -> Self {
         return self.init(buildPlan: buildPlan,
                          graph: graph,
                          logger: inject())
